CIV-5093: Refactor and remove unnecssary code
