+++ src/main/java/uk/gov/hmcts/reform/civil/handler/tasks/CheckStayOrderDeadlineEndTaskHandler.java
@@ -1,22 +1,28 @@
 package uk.gov.hmcts.reform.civil.handler.tasks;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.camunda.bpm.client.task.ExternalTask;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
 import org.springframework.stereotype.Component;
+import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.civil.enums.YesOrNo;
 import uk.gov.hmcts.reform.civil.helpers.CaseDetailsConverter;
 import uk.gov.hmcts.reform.civil.model.CaseData;
+import uk.gov.hmcts.reform.civil.model.genapplication.GAJudicialMakeAnOrder;
 import uk.gov.hmcts.reform.civil.service.CoreCaseDataService;
 import uk.gov.hmcts.reform.civil.service.search.OrderMadeSearchService;
 
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 import static java.time.LocalDate.now;
 import static uk.gov.hmcts.reform.civil.callback.CaseEvent.END_SCHEDULER_CHECK_STAY_ORDER_DEADLINE;
+import static uk.gov.hmcts.reform.civil.callback.CaseEvent.UPDATE_GA_CASE_DATA;
+import static uk.gov.hmcts.reform.civil.enums.CaseState.ORDER_MADE;
 import static uk.gov.hmcts.reform.civil.enums.dq.GeneralApplicationTypes.STAY_THE_CLAIM;
 
 @Slf4j
@@ -30,7 +36,7 @@ public class CheckStayOrderDeadlineEndTaskHandler implements BaseExternalTaskHan
     private final CoreCaseDataService coreCaseDataService;
 
     private final CaseDetailsConverter caseDetailsConverter;
-
+    private final ObjectMapper mapper;
     @Override
     public void handleTask(ExternalTask externalTask) {
         List<CaseData> cases = getOrderMadeCasesThatAreEndingToday();
@@ -53,10 +59,26 @@ public class CheckStayOrderDeadlineEndTaskHandler implements BaseExternalTaskHan
 
     private void fireEventForStateChange(CaseData caseData) {
         Long caseId = caseData.getCcdCaseReference();
-        log.info("Firing event CHECK_STAY_ORDER_END_DATE to check applications with ORDER_MADE"
+        log.info("Firing event END_SCHEDULER_CHECK_STAY_ORDER_DEADLINE to check applications with ORDER_MADE"
                      + "and with Application type Stay claim and its end date is today"
                      + "for caseId: {}", caseId);
 
-        coreCaseDataService.triggerEvent(caseId, END_SCHEDULER_CHECK_STAY_ORDER_DEADLINE);
+        coreCaseDataService.triggerGaEvent(caseId, END_SCHEDULER_CHECK_STAY_ORDER_DEADLINE,
+                                           getUpdatedCaseDataMapper(updateCaseData(caseData)));
+        log.info("Checking state for caseId: {}", caseId);
+    }
+
+    private CaseData updateCaseData(CaseData caseData) {
+        GAJudicialMakeAnOrder judicialDecisionMakeOrder = caseData.getJudicialDecisionMakeOrder();
+        caseData = caseData.toBuilder()
+            .judicialDecisionMakeOrder(
+                judicialDecisionMakeOrder.toBuilder().esOrderProcessedByStayScheduler(YesOrNo.YES).build())
+            .build();
+        return caseData;
+    }
+
+    private Map<String, Object> getUpdatedCaseDataMapper(CaseData caseData) {
+        Map<String, Object> output = caseData.toMap(mapper);
+        return output;
     }
 }
